Load proxy configurations in code whisperer base class
