+++ server/aws-lsp-codewhisperer/src/language-server/utils.ts
@@ -1,10 +1,12 @@
 import { BearerCredentials, CredentialsProvider, Position } from '@aws/language-server-runtimes/server-interface'
-import { AWSError } from 'aws-sdk'
+import { AWSError, ConfigurationOptions } from 'aws-sdk'
 import { distance } from 'fastest-levenshtein'
 import { Suggestion } from './codeWhispererService'
 import { CodewhispererCompletionType } from './telemetry/types'
 import { BUILDER_ID_START_URL, MISSING_BEARER_TOKEN_ERROR } from './constants'
 import { ServerInfo } from '@aws/language-server-runtimes/server-interface/runtime'
+import { readFileSync } from 'fs'
+import { HttpsProxyAgent } from 'hpagent'
 export type SsoConnectionType = 'builderId' | 'identityCenter' | 'none'
 
 export function isAwsError(error: unknown): error is AWSError {
@@ -123,4 +125,25 @@ export function getEndPositionForAcceptedSuggestion(content: string, startPositi
         }
     }
     return endPosition
+}
+
+export const makeProxyConfig = () => {
+    let additionalAwsConfig: ConfigurationOptions = {}
+    const proxyUrl = process.env.HTTPS_PROXY ?? process.env.https_proxy
+
+    if (proxyUrl) {
+        const certs = process.env.AWS_CA_BUNDLE ? [readFileSync(process.env.AWS_CA_BUNDLE)] : undefined
+        const agent = new HttpsProxyAgent({
+            proxy: proxyUrl,
+            ca: certs,
+        })
+
+        additionalAwsConfig = {
+            httpOptions: {
+                agent: agent,
+            },
+        }
+    }
+
+    return additionalAwsConfig
 }
